CC3220: Use built-in flash for SPIFFS
We can drop the SLFS container contraption now! It doesn't work properly on CC3220 anyway: it is no longer possible to read an SLFS file while it's open for writing, it needs to be re-opened read-only, which means even more container switches and even more terrible write performance. Good riddance.

Instead, we put SPIFFS at the end of internal flash. The TRM speaks about flash bing organized in two banks of 512K, so i think this limits code size to 512K (or we'll have to put flash programming routines in RAM and play other tricks to avoid trying to execute code from the bank that is being programmed). But we're nowhere near 512K anyway (~200K currently), so it's grand.

Since it is not possible to program internal flash directly when flashing, we instead put a container on SLFS and transfer it to internal flash on first boot.

TI's flash controller is a bit anal-retentive about writes and raises an error if you try to flip bits from 0 to 1 instead of just ignoring it like everybody else in the world.
Turns out, SPIFFS is playing fast and loose with this in some places, which has been reported upstream: https://github.com/pellepl/spiffs/issues/172
This PR integrates https://github.com/pellepl/spiffs/pull/173 which is my humble attempt at fixing it, and seems to work ok.

PUBLISHED_FROM=65ffe0169b75a5c5685f6a336eb6fd875f181195