ESP32: Fix mode control during scan when ap and sta are disabled
When wifi is diabled but a scan is requested, we bring it up in station
mode for the duration of the scan, then shut down again.

There was a bug here and it would only be done correctly the first time,
after that it would not be started.

Fixes https://github.com/mongoose-os-libs/wifi/issues/39