STA scan and connection rewrite
Implement our own AP selection algorithm, do not rely on SDK.
The algorithm tries to connect to AP with best signal but doesn't get
stuck on one (or a subset) of them. If there is an AP that can be
connected to, it should eventually be found.

Also implement roaming: if device is connected to an AP with RSII less
than wifi.sta_roam_rssi_thr, it will periodically
(wifi.sta_roam_interval) scan for a better AP and try to reconnect.
AP must be at least 5 dbm better than the one we are connected to.
Roaming is not seamless, i.e. connection is lost during attempt.
Scanning also introduces blips, so if connection stability is more
important, roaming should not be used.