Add quick connect mode
Use `bssid` and `channel` fields of the config to store BSSID and channel of
the last AP successfulyl connected to.
These are stored with `*` prefix and as a negated value to distinguish
them from user-provided values.
When first connecting, check all configs for entries with BSSID and
channel specified and build initial conenct queue with these static
entries, bypassing the scan. If no successful connection is made, we go
back to scanning and building a new queue as before.

Signed-off-by: Deomid "rojer" Ryabkov <rojer@rojer.me>